Adding vendor/merchant model interactions
